requirements.txt
# Thư viện cơ bản
Flask==2.3.3
Flask-Mail==0.9.1
#--------------------
gunicorn==21.2.0
#------------------------
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Cors==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7

# Database
#------------------------------
psycopg2-binary==2.9.9
#--------------------------------
cloudinary
#--------------------------------
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.5

# QR Code
qrcode==7.4.2
Pillow==10.1.0

# Blockchain
web3==6.11.1
py-solc-x==1.1.1

# Utilities
requests==2.31.0
pydantic==2.5.2
email-validator==2.1.0.post1
python-dateutil==2.8.2
pytz==2023.3.post1

# Development
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2
black==23.11.0
flake8==6.1.0

# OpenAI
openai==1.3.0
